First DD-WRT forum post
Source: http://www.dd-wrt.com/phpBB2/viewtopic.php?t=44324
Post Author: frater
Post Date: Dec 21, 2008 11:35

When DD-WRT does a backup of the settings this is done in a binary way. Even if no variables have changed and the hardware has not changed, it is not totally safe to restore this binary backup on another version.

I'm saving these variables to a script. This was not my idea, but the other implementation on this forum was saving more than it should.

This script also creates some variables that don't really exist. I previously wrote a script that was able to avoid this, but it had a terrible speed-impact. The variables it creates (makes up) comes from data containing a "=". This really is not a problem.

I'm deliberately skipping some variables that contain hardware-specific data. This way it is safe to restore these settings on different hardware.
After an upgrade you can run this restore-script and be up and running in no time.