Add support for vars_preferred
Source: http://www.dd-wrt.com/phpBB2/viewtopic.php?p=291362#291362
Post Author: frater
Post Date: Apr 22, 2009 13:28

An include-file was what I first wrote, but this means you're missing out on variables that are getting introduced in the future. But if one, like me, is doing a regular compare this will not get unnoticed.

Someone in this thread prefers code that's coming directly from me and because I don't want to support someone else's code I have written a "include" as well.

So.. here it is... test away