Add -r to sort command
Source: http://www.dd-wrt.com/phpBB2/viewtopic.php?p=328013#328013
Post Author: frater
Post Date: Jul 26, 2009

Thanks hardwarewizard for adding the comments, finding the "feature" in DD-WRT and adding PREFERRED to that echo line (cosmetic).

Here's the script that should take care of the order in which the variables are written.